refactor: move duplicate code into hook
