+++ packages/tradershub/src/features/cfd/modals/ButtonGroups/AddAccountButtonsGroup.tsx
@@ -1,9 +1,7 @@
 import React from 'react';
 import { ButtonGroup } from '@/components';
-import { useQueryParams } from '@/hooks';
+import { useHandleSendEmail, useQueryParams } from '@/hooks';
 import { useCFDContext } from '@/providers';
-import { platformPasswordResetRedirectLink } from '@/utils';
-import { useActiveTradingAccount, useSettings, useVerifyEmail } from '@deriv/api-v2';
 import { Button } from '@deriv-com/ui';
 import { CFDPlatforms, PlatformDetails } from '../../constants';
 import DxtradeCreateAccountButton from '../DxtradePasswordModal/DxtradeCreateAccountButton';
@@ -15,31 +13,18 @@ type TAddAccountButtonsGroupProps = {
 
 const AddAccountButtonsGroup = ({ password }: TAddAccountButtonsGroupProps) => {
     const { cfdState } = useCFDContext();
-    const { data } = useSettings();
-    const { mutate } = useVerifyEmail();
-    const { data: activeTrading } = useActiveTradingAccount();
     const { platform = CFDPlatforms.MT5 } = cfdState;
     const { openModal } = useQueryParams();
+    const { handleSendEmail } = useHandleSendEmail();
 
-    const handleSendEmail = async () => {
-        if (data.email) {
-            await mutate({
-                type:
-                    platform === PlatformDetails.dxtrade.platform
-                        ? 'trading_platform_dxtrade_password_reset'
-                        : 'trading_platform_mt5_password_reset',
-                url_parameters: {
-                    redirect_to: platformPasswordResetRedirectLink(platform, activeTrading?.is_virtual),
-                },
-                verify_email: data.email,
-            });
-        }
+    const handleForgotPassword = async () => {
+        handleSendEmail();
         openModal('SentEmailContentModal');
     };
 
     return (
         <ButtonGroup className='justify-end w-full'>
-            <Button color='black' onClick={handleSendEmail} variant='outlined'>
+            <Button color='black' onClick={handleForgotPassword} variant='outlined'>
                 Forgot password?
             </Button>
             {platform === PlatformDetails.dxtrade.platform && (
